requirements.txt
numpy>=2.0
scipy>=1.13
matplotlib>=3.8
cvxpy>=1.5
qiskit>=2.0
qiskit-aer>=0.15
qiskit-algorithms>=0.3
qiskit-ibm-runtime>=0.30
pennylane>=0.39
pylatexenc>=2.10

# Backend
fastapi>=0.115
uvicorn[standard]>=0.32
python-dotenv>=1.0
python-multipart>=0.0.12
pydantic>=2.9
sse-starlette>=2.1

# Testing
pytest>=8.0
httpx>=0.27
